user_service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["NiLNight <bkmz-2020@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
alembic = "^1.17.2"
email-validator = "^2.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.21"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
